requirements.txt
Django==4.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.1
djangorestframework-simplejwt==5.3.1
requests==2.31.0
Pillow==10.1.0
django-cors-headers==4.0.0
fastapi==0.110.0
uvicorn==0.29.0
PyPDF2==3.0.1
pypdf==4.2.0
langchain==0.1.16
chromadb==0.4.24
langchain-community==0.0.30
